+++ Source/SwiftLintFramework/Rules/MissingDocsRule.swift
@@ -55,25 +55,41 @@ extension File {
 
 public enum AccessControlLevel: String, CustomStringConvertible {
     case Private = "source.lang.swift.accessibility.private"
+    case FilePrivate = "source.lang.swift.accessibility.fileprivate"
     case Internal = "source.lang.swift.accessibility.internal"
     case Public = "source.lang.swift.accessibility.public"
+    case Open = "source.lang.swift.accessibility.open"
 
     internal init?(description value: String) {
         switch value {
         case "private": self = .Private
+        case "fileprivate": self = .FilePrivate
         case "internal": self = .Internal
         case "public": self = .Public
+        case "open": self = .Open
         default: return nil
         }
+    }
+
+    init?(identifier value: String) {
+        self.init(rawValue: value)
     }
 
     public var description: String {
         switch self {
         case Private: return "private"
+        case FilePrivate: return "fileprivate"
         case Internal: return "internal"
         case Public: return "public"
+        case Open: return "open"
         }
     }
+
+    // Returns true if is `private` or `fileprivate`
+    var isPrivate: Bool {
+        return self == .Private || self == .FilePrivate
+    }
+
 }
 
 public struct MissingDocsRule: OptInRule {
@@ -92,7 +108,8 @@ public struct MissingDocsRule: OptInRule {
     }
 
     public init() {
-        parameters = [RuleParameter(severity: .Warning, value: .Public)]
+        parameters = [RuleParameter(severity: .Warning, value: .Public),
+                      RuleParameter(severity: .Warning, value: .Open)]
     }
 
     public let parameters: [RuleParameter<AccessControlLevel>]